refactor init into history and sessions
